separate recording and parsing
