+++ .gitignore
@@ -134,3 +134,5 @@ dmypy.json
 
 # Streamlit
 *secrets.toml*
+
+logs/*
